+++ src/main/java/calculator/view/OutputView.java
@@ -4,7 +4,7 @@ public class OutputView {
 
     private static final String RESULT_MESSAGE = "결과 : ";
 
-    public void viewResult(int sum) {
+    public void viewResult(long sum) {
         System.out.print(RESULT_MESSAGE + sum);
     }
 }
