+++ src/test/java/calculator/StringAdderTest.java
@@ -25,7 +25,7 @@ public class StringAdderTest {
         this.calculator = new Calculator(validator, parser, extractor);
     }
 
-    private int calculate(String inputString) {
+    private long calculate(String inputString) {
         return calculator.calculate(inputString);
     }
 
@@ -33,7 +33,7 @@ public class StringAdderTest {
     @DisplayName("단일 숫자 문자열")
     void singleNumberTest() {
         String inputString = "5";
-        int sum = calculate(inputString);
+        long sum = calculate(inputString);
         assertThat(sum).isEqualTo(5);
     }
 
@@ -41,7 +41,7 @@ public class StringAdderTest {
     @DisplayName("쉼표로 구분된 숫자")
     void commaSeparatedTest() {
         String inputString = "1,2,3";
-        int sum = calculate(inputString);
+        long sum = calculate(inputString);
         assertThat(sum).isEqualTo(6);
     }
 
@@ -49,7 +49,7 @@ public class StringAdderTest {
     @DisplayName("콜론으로 구분된 숫자")
     void colonSeparatedTest() {
         String inputString = "1:2:3:4";
-        int sum = calculate(inputString);
+        long sum = calculate(inputString);
         assertThat(sum).isEqualTo(10);
     }
 
@@ -57,7 +57,7 @@ public class StringAdderTest {
     @DisplayName("쉼표와 콜론 혼합")
     void mixedDelimiterTest() {
         String inputString = "1,2:3";
-        int sum = calculate(inputString);
+        long sum = calculate(inputString);
         assertThat(sum).isEqualTo(6);
     }
 
@@ -65,7 +65,7 @@ public class StringAdderTest {
     @DisplayName("커스텀 구분자")
     void customSemicolonTest() {
         String inputString = "//;\\n1;2;3";
-        int sum = calculate(inputString);
+        long sum = calculate(inputString);
         assertThat(sum).isEqualTo(6);
     }
 
@@ -73,7 +73,7 @@ public class StringAdderTest {
     @DisplayName("커스텀 구분자 파이프")
     void customPipeTest() {
         String inputString = "//|\\n4|5|6";
-        int sum = calculate(inputString);
+        long sum = calculate(inputString);
         assertThat(sum).isEqualTo(15);
     }
 
@@ -81,7 +81,7 @@ public class StringAdderTest {
     @DisplayName("단일 숫자 + 커스텀 구분자")
     void singleNumberCustomDelimiterTest() {
         String inputString = "//;\\n7";
-        int sum = calculate(inputString);
+        long sum = calculate(inputString);
         assertThat(sum).isEqualTo(7);
     }
 
@@ -89,7 +89,7 @@ public class StringAdderTest {
     @DisplayName("특수문자 커스텀 구분자")
     void specialCharacterDelimiterTest() {
         String inputString = "//^\\n1^2^3";
-        int sum = calculate(inputString);
+        long sum = calculate(inputString);
         assertThat(sum).isEqualTo(6);
     }
 }
